Add tests for 404 when updating or deleting a user that does not exist
The `PUT` and `DELETE` endpoints for updating and deleting users were previously tested to ensure that they correctly handled the case where the user being updated or deleted does not exist. However, the tests did not cover the case where the user being updated or deleted is not found.

To ensure that these endpoints behave correctly in this case, I have added tests that verify that a 404 status code is returned and that the correct error message is included in the response body.